requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.3
sqlalchemy[asyncio]==2.0.41
asyncpg==0.30.0
alembic==1.15.2
python-jose[cryptography]==3.4.0
bcrypt==4.3.0
pydantic-settings==2.9.1
python-multipart==0.0.20
